refactor : rename follow folder name into social
